requirements.txt
pillow
pytz
requests
flask
RGBMatrixEmulator
systemd-python; sys_platform == "linux" # only install on Linux
scipy
numpy
types-requests
types-pytz
types-Pillow
scipy-stubs
types-RPi.GPIO